More numpy types (#5683)
12 errors left

Part of https://github.com/quantumlib/Cirq/issues/3767

+++ cirq-core/cirq/ops/pauli_string.py
@@ -744,7 +744,8 @@ class PauliString(raw_types.Operation, Generic[TKey]):
 
         while any(result.shape):
             result = np.trace(result, axis1=0, axis2=len(result.shape) // 2)
-        return result * self.coefficient
+
+        return float(result * self.coefficient)
 
     def zip_items(
         self, other: 'cirq.PauliString[TKey]'
